added test for new method
